+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Update to dropwizard 3.0.16
 * Use guice version without bundled asm ('classes' classifier)
+* Support field injections in application (to use injected services in run method)
 * Customizable DefaultTestClientFactory: it is now possible to use default implementation
   with customizations (override `configure` method)
     - Add ApacheTestClientFactory: useful to support PATCH methods on jdk > 16